Allow to automatically sync DTBs when /boot is on a separate partition (Closes: #1025956)
Having `/boot` on a separate partition is a very common use-case,
requiring device tree binary files to be copied over to this partition
as `u-boot` won't be searching other partitions for those.

Up until now, users had to manually copy (and potentially edit) the
provided `zz-sync-dtb` example script if their system was using a
separate `/boot` partition, which isn't practical.

This patch implements a way to automate this synchronization process
in such cases, and remove the synchronized files when the corresponding
kernel is removed as well.

In order to maintain the current behaviour, this feature is disabled by
default and must be enabled by setting the `U_BOOT_SYNC_DTBS`
configuration variable to `true`.